change default run coordinator
